Add modal
I created an article with a class modal-container, inside I created a div with a class of modal.
Then, I made an h1 a ul for the languages and a p. After, I created two wrappers for the buttons and the buttons.

I styled the modal with a mobile first approach.